requirements.txt
flask==3.0.3
gunicorn==23.0.0
ultralytics==8.3.12
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
torch==2.3.0
torchvision==0.18.0
